feat: create dev-ci image embedding the frontend
