+++ .gitignore
@@ -21,7 +21,8 @@ mysql-*.tar*
 .air
 
 # Frontend asset
-dist
+server/dist/*
+!server/dist/index.html
 
 # Dev database
 data
@@ -34,5 +35,4 @@ bytebase-build
 
 .DS_Store
 
-dist/
 docs/openapi/swagger.json
